Moved the copy logic to the FormProgressBar.
